+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+
+## 6.0.0
 * Changed `filterMatcher` parameters: it will now receive rowData and a filterValue. This is a breaking change.
 
 ## 5.7.2